Split the single web socket test into multiple.
